+++ src/models/groupSettings/groupSettings.ts
@@ -1,0 +1,3 @@
+import { ObjectSettings } from '../objectSettings';
+
+export abstract class GroupSettings extends ObjectSettings {}